+++ Content.Client/UserInterface/Systems/Inventory/InventoryUIController.cs
@@ -357,7 +357,7 @@ public sealed partial class InventoryUIController : UIController, IOnStateEntere
                 if (!slot.InsertOnInteract)
                     continue;
 
-                if (!itemSlotsSys.CanInsert(container.ContainedEntity.Value, held.Value, null, slot))
+                if (!itemSlotsSys.CanInsert(container.ContainedEntity.Value, slot, held.Value, null))
                     continue;
                 fits = true;
                 break;
