+++ Content.Shared/Containers/ItemSlots/ItemSlotsLockComponent.cs
@@ -1,3 +1,4 @@
+using Content.Shared.Lock;
 using Robust.Shared.GameStates;
 
 namespace Content.Shared.Containers.ItemSlots;
