+++ Content.Shared/Kitchen/EntitySystems/SharedReagentGrinderSystem.cs
@@ -237,7 +237,7 @@ public abstract partial class SharedReagentGrinderSystem : EntitySystem
         if (!_power.IsPowered(ent.Owner))
             return;
 
-        var beaker = _itemSlotsSystem.GetItemOrNull(ent, ReagentGrinderComponent.BeakerSlotId);
+        var beaker = _itemSlotsSystem.GetItemOrNull(ent.Owner, ReagentGrinderComponent.BeakerSlotId);
 
         // Do we have anything to grind/juice and a container to put the reagents in?
         if (ent.Comp.InputContainer.ContainedEntities.Count <= 0 || !HasComp<FitsInDispenserComponent>(beaker))
