+++ Content.Shared/Medical/Cryogenics/SharedCryoPodSystem.cs
@@ -343,7 +343,7 @@ public abstract partial class SharedCryoPodSystem : EntitySystem
         if (patient == null)
             return; // Refuse to inject if there is no patient.
 
-        var beaker = _itemSlots.GetItemOrNull(cryoPod, cryoPod.Comp.SolutionContainerName);
+        var beaker = _itemSlots.GetItemOrNull(cryoPod.Owner, cryoPod.Comp.SolutionContainerName);
 
         if (beaker == null
             || !beaker.Value.Valid
@@ -387,9 +387,8 @@ public abstract partial class SharedCryoPodSystem : EntitySystem
             return (null, null);
 
         var beaker = _itemSlots.GetItemOrNull(
-            entity.Owner,
-            entity.Comp.SolutionContainerName,
-            itemSlotsComponent
+            (entity.Owner, itemSlotsComponent),
+            entity.Comp.SolutionContainerName
         );
 
         if (beaker == null
