+++ Content.Shared/Silicons/StationAi/SharedStationAiSystem.cs
@@ -235,7 +235,7 @@ public abstract partial class SharedStationAiSystem : EntitySystem
             return;
 
         // Try to insert our thing into them
-        if (_slots.CanEject(ent.Owner, args.User, ent.Comp.Slot))
+        if (_slots.CanEject(ent.Owner, ent.Comp.Slot, args.User))
         {
             if (!_slots.TryInsert(args.Args.Target.Value, targetHolder.Slot, ent.Comp.Slot.Item!.Value, args.User, excludeUserAudio: true))
             {
@@ -247,7 +247,7 @@ public abstract partial class SharedStationAiSystem : EntitySystem
         }
 
         // Otherwise try to take from them
-        if (_slots.CanEject(args.Args.Target.Value, args.User, targetHolder.Slot))
+        if (_slots.CanEject(args.Args.Target.Value, targetHolder.Slot, args.User))
         {
             if (!_slots.TryInsert(ent.Owner, ent.Comp.Slot, targetHolder.Slot.Item!.Value, args.User, excludeUserAudio: true))
             {
