+++ Content.Shared/Weapons/Ranged/Systems/SharedGunSystem.ChamberMagazine.cs
@@ -163,7 +163,7 @@ public abstract partial class SharedGunSystem
 
             if (slots != null)
             {
-                _slots.SetLock(uid, ChamberSlot, true, slots);
+                _slots.SetLock((uid, slots), ChamberSlot, true);
             }
 
             Audio.PlayPredicted(component.BoltClosedSound, uid, user);
@@ -193,7 +193,7 @@ public abstract partial class SharedGunSystem
 
             if (slots != null)
             {
-                _slots.SetLock(uid, ChamberSlot, false, slots);
+                _slots.SetLock((uid, slots), ChamberSlot, false);
             }
 
             Audio.PlayPredicted(component.BoltOpenedSound, uid, user);
